Changed logic LineFollow, added rotation counter
Change the LineFollow.nxc program so that it will keep powering one
motor until the opposite sensor detects tape, which will keep powering
it's motor until the other sensor detects... and so on.

The Rotation count program is designed to calculate the amount one motor
has turned, which was use to verify that it counts to high integers
(>10000), enough to record a 'macro'. Certain parts of this program will
be implemented into our LineFollow later (if not totally changing the
logic)